Change Movie.totalVotes from double to int

+++ include/Movie.h
@@ -20,14 +20,14 @@ class Movie {
         int releaseYear;
 
     public:
-        Movie(string name, string genre, string director, string starringActor, double rating, double totalVotes, int releaseYear);
+        Movie(string name, string genre, string director, string starringActor, double rating, int totalVotes, int releaseYear);
         string getName();
         string getGenre();
         string getDirector();
         int getReleaseYear();
         string getStarringActor();
         double getRating();
-        double getTotalVotes();
+        int getTotalVotes();
 
 
 };
